Move `License` out of `views` into the base model package
