+++ src/main/java/ch/usi/si/seart/model/License.java
@@ -1,0 +1,43 @@
+package ch.usi.si.seart.model;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.experimental.FieldDefaults;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
+import javax.validation.constraints.NotBlank;
+import java.util.HashSet;
+import java.util.Set;
+
+@FieldDefaults(level = AccessLevel.PRIVATE)
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "license")
+@Entity
+public class License {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "id")
+    Long id;
+
+    @NotBlank
+    @Column(name = "name", unique = true)
+    String name;
+
+    @Builder.Default
+    @OneToMany(mappedBy = "license")
+    Set<GitRepo> repos = new HashSet<>();
+}
