--- src/main/java/ch/usi/si/seart/model/view/License.java
@@ -1,25 +1,0 @@
-package ch.usi.si.seart.model.view;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.experimental.FieldDefaults;
-import org.hibernate.annotations.Immutable;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@FieldDefaults(level = AccessLevel.PRIVATE)
-@Entity
-@Immutable
-@Table(name = "license")
-public class License {
-
-    @Id
-    String name;
-}
